% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAKERNAS.R
\name{SAKERNAS}
\alias{SAKERNAS}
\title{Create Survei Angkatan Kerja Nasional (SAKERNAS)}
\usage{
SAKERNAS(data)
}
\arguments{
\item{data}{dataframes}
}
\value{
A dataframes of Survei Angkatan Kerja Nasional (SAKERNAS) of Indonesia
}
\description{
Create Survei Angkatan Kerja Nasional (SAKERNAS)
}
