% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAKERNAS2020.R
\docType{data}
\name{SAKERNAS2020}
\alias{SAKERNAS2020}
\title{SAKERNAS 2020}
\format{
A data frame with 132 variables:
\describe{
\item{TAHUN}{Tahun}
\item{URUTAN}{Urutan}
\item{SMT}{Semester}
\item{FINAL_WEIG}{Penimbang Official}
\item{KODE_PROV}{Kode Provinsi}
\item{NAMA_PROV}{Nama Provinsi}
\item{KODE_KAB}{Kode kabupaten}
\item{NAMA_KAB}{Nama Kabupaten}
\item{ID_NKS}{ID NKS}
\item{NO_DSRT}{No Urut Sampel}
\item{KLASIFIKAS}{Klasifikasi Desa/Kelurahan}
\item{JLH_ART}{jumlah ART}
\item{JLH_ART5}{jumlah ART usia 5 tahun keatas}
\item{K1}{No urut}
\item{K3}{Hubungan dgn KRT}
\item{K4}{Jenis Kelamin}
\item{K5_BL}{Bulan lahir}
\item{K5_TH}{Tahun lahir}
\item{K6}{Umur}
\item{R4}{Status Perkawinan}
\item{R5}{Partisipasi sekolah}
\item{R6A}{6a. Ijazah tertinggi}
\item{R6B}{6b. Jurusan pendidikan}
\item{R6C}{6c. Lulus dalam setahun terakhir}
\item{R6D}{6d. Pernah kursus & mendapat sertifikat}
\item{R6E}{6e. Sedang mengikuti pelatihan/kursus/training (tdk harus bersertifikat)}
\item{R7}{7. Pada Agustus 2015 dimana tempat tinggal anda}
\item{R7_PROV}{7a. Provinsi tinggal 5th lalu}
\item{R7_KAB}{7b. Kab/kota tinggal 5th lalu}
\item{R8A}{8a. Kesulitan melihat}
\item{R8B}{8b. Kesulitan mendengar}
\item{R8C}{8c. Kesulitan berjalan}
\item{R8D}{8d. Kesulitan memegang}
\item{R8E}{8e. Kesulitan berbicara}
\item{R8F}{8f. Kesulitan lain}
\item{R9A}{9a. Bekerja minimal 1 Jam}
\item{R9B}{9b. Melakukan kegiatan utk memperoleh penghasilan/pendapatan uang}
\item{R9C}{9c. Membantu kegiatan usaha atau pekerjaan keluarga/orang lain}
\item{R10A}{10a. Sementara tidak bekerja seminggu yang lalu}
\item{R10B}{10b. Alasan sementara tidak bekerja}
\item{R10C}{10c. Apakah alasan terkait dengan covid}
\item{R10D}{10d. Jaminan kembali bekerja pada unit usaha/tempat kerja sekarang}
\item{R12A}{12a. Apakah Status pekerjaan anda}
\item{R12B}{12b. Berapa jumlah buruh/karyawan/pegawai yang dibayar}
\item{R13A_KATEG}{Lapangan Usaha 17 Kategori}
\item{R13B_KJI19}{Klasifikasi Baku Jenis Pekerjaan 1982 - 1 digit}
\item{R13B_KBJI2}{KBJI 2014}
\item{R14A1}{14a1. Gaji(uang)}
\item{R14A2}{14a2. Gaji(barang)}
\item{R14B}{14b. Apakah terjadi Perubahan penghasilan}
\item{R15A1}{15a. Bulan mulai bekerja}
\item{R15A2}{15a. Tahun mulai bekerja}
\item{R15B}{15b. Lama mencari pekerjaan/mempersiapkan usaha}
\item{R16A1}{16a. Jam kerja senin}
\item{R16A2}{16a. Jam kerja selasa}
\item{R16A3}{16a. Jam kerja rabu}
\item{R16A4}{16a. Jam kerja kamis}
\item{R16A5}{16a. Jam kerja jumat}
\item{R16A6}{16a. Jam kerja sabtu}
\item{R16A7}{16a. Jam kerja minggu}
\item{R16A}{16a. Jumlah jam kerja seminggu (pembulatan statistik)}
\item{R16B}{16b. Apakah terjadi perubahan jam kerja}
\item{R16C}{16c. Alasan terjadi perubahan Jam kerja}
\item{R16D}{16d. Alasan terkait dengan CORONA/COVID-19}
\item{R17A}{17a. Menggunakan internet}
\item{R17B1}{17b1. Pemanfaatan internet untuk promosi}
\item{R17B2}{17b2. Pemanfaatan internet untuk komunikasi}
\item{R17B3}{17b3. Pemanfaatan internet untuk transaksi}
\item{R18A}{18a. Menerapkan kegiatan bekerja dari rumah (WFH)}
\item{R18B}{18b. Kendala utama  kegiatan bekerja dari rumah (WFH)}
\item{R18C}{18c. Bekerja di tempat keramaian/kerumunan}
\item{R18D1}{18d1. Lokasi kerja Di Rumah Sendiri}
\item{R18D2}{18d2. Lokasi kerja Di Pasar}
\item{R18D3}{18d3. Lokasi kerja Di Bioskop}
\item{R18D4}{18d4. Lokasi kerja Di Mall/Ruko}
\item{R18D5}{18d5. Lokasi kerja Di Terminal}
\item{R18D6}{18d6. Lokasi kerja Di Pinggir Jalan}
\item{R18D7}{18d7. Lokasi kerja Lainnya}
\item{R19A}{19a. Seminggu yang lalu Bekerja di rumah sendiri}
\item{R19B}{19b. Dimana Lokasi Tempat Kerja}
\item{R19B_PROV}{19b. Prov tempat kerja}
\item{R19B_KAB}{19b. Kab tempat kerja}
\item{R19C}{19c. Kerja pulang pergi}
\item{R19D}{19d. Jenis transportasi}
\item{R20A}{20a. Memiliki pekerjaan tambahan}
\item{R20B1}{20b. Jam kerja senin}
\item{R20B2}{20b. Jam kerja selasa}
\item{R20B3}{20b. Jam kerja rabu}
\item{R20B4}{20b. Jam kerja kamis}
\item{R20B5}{20b. Jam kerja jumat}
\item{R20B6}{20b. Jam kerja sabtu}
\item{R20B7}{20b. Jam kerja minggu}
\item{R20B}{20b. Jumlah jam kerja satu minggu (pembulatan Statistik)}
\item{R21A}{21a. Menambah jam kerja}
\item{R21B}{21b. Alasan tidak ingin menambah jam kerja}
\item{R21C}{21c. Alasan terkait dengan CORONA/COVID-19}
\item{R22A}{22a. Mulai mencari pekerjaan/usaha}
\item{R22B}{22b. Mempersiapkan suatu kegiatan usaha baru}
\item{R23A}{23. Lama mencari pekerjaan/mempersiapkan usaha (tahun)}
\item{R23B}{23. Lama mencari pekerjaan/mempersiapkan usaha (bulan)}
\item{R24A}{24a. Mendaftar bursa kerja}
\item{R24B}{24b. Menghubungi perusahaan/mencari pekerjaan di media cetak/elektonik/internet}
\item{R24C}{24c. Mengiklankan diri di media cetak/elektronik/internet}
\item{R24D}{24d. Menghubungi keluarga/kenalan}
\item{R24E}{24e. Mengumpulkan modal/mencari lokasi/ mengurus surat izin usaha}
\item{R24F}{24f. Alasan lain}
\item{R25A}{25a. Alasan tidak mencari pekerjaan/mempersiapkan usaha seminggu yang lalu}
\item{R25B}{25b. Alasan terkait dengan CORONA/COVID-19}
\item{R26}{26. Menerima tawaran pekerjaan}
\item{R27A}{27a. Mengetahui program kartu prakerja}
\item{R27B}{27b. Mendaftar program kartu prakerja}
\item{R27C}{27c. Alasan Mendaftar program kartu prakerja}
\item{R27D}{27d. Lolos seleksi program kartu prakerja}
\item{R27E}{27e. Menyelesaikan pelatihan pada program kartu prakerja}
\item{R27F}{27f. Program kartu prakerja Meningkatkan keterampilan kerja}
\item{R27G}{27g. Mendapatkan uang saku (insentif) dari Program kartu prakerja Meningkatkan keterampilan kerja}
\item{R27H1}{27h1. Memenuhi Kebutuhan sehari-hari}
\item{R27H2}{27h2. Modal Usaha}
\item{R27H3}{27h3. Membayar hutang}
\item{R27H4}{27h4. Ditabung}
\item{R27H5}{27h5. Lainnya}
\item{R27I}{27i. Program yang dibutuhkan terkait dengan CORONA/COVID-19}
\item{R28}{28. Pernah mempunyai pekerjaan/usaha sebelumnya}
\item{R29A}{29a. Pernah berhenti kerja periode  Agustus 2019-Agustus 2020}
\item{R29B_BL}{29b. Bulan berhenti bekerja}
\item{R29B_TH}{29b. Tahun berhenti bekerja}
\item{R30A}{30a. Alasan utama berhenti bekerja}
\item{R30B}{30b. Alasan terkait dengan CORONA/COVID-19}
\item{R31A}{31a. Dalam Seminggu terakhir bersekolah}
\item{R31B}{31b. Dalam Seminggu terakhir mengurus rumah tangga}
\item{R31C}{31c. Dalam Seminggu terakhir melakukan kegiatan lainnya}
\item{R31D}{31d. Waktu terbanyak dalam Seminggu terakhir melakukan kegiatan}
}
}
\source{
\describe{
https://www.bps.go.id
}
}
\usage{
SAKERNAS2020
}
\description{
SAKERNAS 2020
}
\examples{
data(SAKERNAS2020)
}
\keyword{datasets}
